+++ .dockerignore
@@ -1,6 +1,6 @@
 *
 
-!botcore/
+!pydis_core/
 !docs/
 !tests/
