+++ pydis_core/utils/_extensions.py
