+++ pydis_core/utils/function.py
