+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "bot-core"
-version = "8.2.1"
-description = "PyDis bot core provides core functionality and utility to the bots of the Python Discord community."
+name = "pydis_core"
+version = "9.0.0"
+description = "PyDis core provides core functionality and utility to the bots of the Python Discord community."
 authors = ["Python Discord <info@pythondiscord.com>"]
 license = "MIT"
 classifiers=[
@@ -12,11 +12,12 @@ classifiers=[
     "Programming Language :: Python :: 3",
 ]
 packages = [
-    { include = "botcore" },
+    { include = "pydis_core" },
 ]
 include = ["LICENSE"]
 exclude = ["tests", "tests.*"]
 readme = "README.md"
+homepage = "https://pythondiscord.com/"
 documentation = "https://bot-core.pythondiscord.com/"
 repository = "https://github.com/python-discord/bot-core"
 keywords = ["bot", "discord", "discord.py"]
@@ -68,5 +69,5 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.coverage.run]
 branch = true
-source_pkgs = ["botcore"]
+source_pkgs = ["pydis_core"]
 source = ["tests"]
