+++ tox.ini
@@ -2,7 +2,7 @@
 max-line-length=120
 docstring-convention=all
 import-order-style=pycharm
-application_import_names=botcore,docs,tests
+application_import_names=pydis_core,docs,tests
 exclude=.cache,.venv,.git,constants.py,bot/
 ignore=
     B311,W503,E226,S311,T000,E731
